todo double

+++ TAREAS/8/main.c
@@ -5,16 +5,16 @@
 int main(int argc, char *argv[]) {
   // x es lo que se quiere sacar
   //y el valor a convertir
-int x= atoi(argv[1]);
-int y = atoi(argv[2]);
+double x= atoi(argv[1]);
+double y = atoi(argv[2]);
 if(x!=0){
   //este pasa de fahrenhait a celsius
-float c = (y -32)/1.8;
+double c = (y -32)/1.8;
 printf("%.2lf",c);
 }
   else{
     // este es de celsius a fahrenhait
-    float f = (y*1.8)+ 32;
+    double f = (y*1.8)+ 32;
     printf("%.2lf",f);
 
   }
